name bugreport changed to contactUs and some minor bug fixes
