--- bugReport/admin.py
@@ -1,4 +1,0 @@
-from django.contrib import admin
-from .models import BugReport
-
-admin.site.register(BugReport)
