+++ contactUs/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class BugreportConfig(AppConfig):
+class ContactusConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'bugReport'
+    name = 'contactUs'
