+++ contactUs/templates/cu_index.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 
 {% block title %}
-Bug Reports
+Contact Us
 {% endblock %}
 
 {% load static %}
@@ -11,7 +11,7 @@ Bug Reports
 {% endblock %}
 
 {% block page_header %}
-Bug Reports
+Contact Us
 {% endblock %}
 
 {% block body %}
@@ -29,7 +29,7 @@ Bug Reports
             </div>
             <div class="form-group" style="margin-bottom: 25px;">
                 <label class="form-label">Description</label>
-                <textarea class="form-control" id="exampleFormControlTextarea1" rows="3" placeholder="Description"
+                <textarea class="form-control" id="exampleFormControlTextarea1" rows="6" placeholder="Description"
                     name="description"></textarea>
             </div>
             <div class="btn-list text-right" style="margin-bottom: 10px;">
