+++ templates/base.html
@@ -157,11 +157,11 @@
                     <li>
                         <h3>Main</h3>
                     </li>
-                    <!-- <li>
-                        <a class="side-menu__item" href="widgets.html"><i
-                                class="side-menu__icon fe fe-codepen"></i><span
-                                class="side-menu__label">Widgets</span></a>
-                    </li> -->
+                    <li>
+                        <a class="side-menu__item" href="/"><i
+                                class="side-menu__icon fe fe-airplay"></i><span
+                                class="side-menu__label">Dashboard</span></a>
+                    </li>
                     <li class="slide is-expanded">
                         <a class="side-menu__item" data-toggle="slide" href="#"><i
                                 class="side-menu__icon fe fe-airplay"></i><span class="side-menu__label">All</span><i
@@ -192,9 +192,9 @@
                         </ul>
                     </li>
                     <li>
-                        <a class="side-menu__item" href="/bugreport"><i
+                        <a class="side-menu__item" href="/contactus"><i
                                 class="side-menu__icon fe fe-file-text"></i><span
-                                class="side-menu__label">Bug Reports</span></a>
+                                class="side-menu__label">Contact Us</span></a>
                     </li>
                     <!-- <li>
                         <a class="side-menu__item" href="widgets.html"><i
